Parsers: don't use data_to_string for non-0-terminated strings
App used `data_to_string()` because it's a fixed sized string with termination
marker byte 0.

Bye was incorrectly using this function as the reason could be an utf-8 string.

This commit removes `data_to_string()` so as to avoid confusion and fixes Bye.
Also adds a string accessor for the SDES Item value.