sdes: SdesChunkBuilder::add_item_owned() doesn't need 'static bound
SdesChunkBuilder::add_item_owned() allows adding an item, possibly with a
shorter lifetime than the SdesChunkBuilder by owning it.

In previous implementation, this function was defined under the
SdesChunkBuilder<'static> impl block. As the function doesn't change the
lifetime of the SdesChunkBuilder, it can actually be defined under the
SdesChunkBuilder<'a> impl block, thus allowing it to be used in less
constrained use cases.